mattel-web/ecomm/core/src/main/java/com/mattel/ecomm/core/pojos/LeftNavLinkPojo.java
package com.mattel.ecomm.core.pojos;

public class LeftNavLinkPojo {
	private String pageName;
	private String pageLink;
	private boolean currentPage = false;

	public String getPageName() {
		return pageName;
	}

	public void setPageName(String pageName) {
		this.pageName = pageName;
	}

	public String getPageLink() {
		return pageLink;
	}

	public void setPageLink(String pageLink) {
		this.pageLink = pageLink;
	}

	public boolean isCurrentPage() {
		return currentPage;
	}

	public void setCurrentPage(boolean currentPage) {
		this.currentPage = currentPage;
	}
}
